pyproject.toml
[tool.poetry]
name = "construction-cost-bot"
version = "1.0.0"
description = "Chatbot de IA para estimacion de costos de construccion"
authors = ["Julio Diaz"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-mcp-adapters = "^0.1.7"
langgraph = "^0.4.8"
langsmith = "^0.1.17"
openai = "^1.14.0"
gradio = "^4.13.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.0"
pytest-mock = "^3.12.0"
black = "^24.0.0"
ruff = "^0.4.0"
mypy = "^1.10.0"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
